+++ include/cpfusa/fusa.hpp
@@ -6,11 +6,11 @@
 
 namespace cpfusa {
 
-constexpr std::string_view Version      = "0.14.0";
+constexpr std::string_view Version      = "0.14.1";
 constexpr std::string_view VersionMajor = "0";
 constexpr std::string_view VersionMinor = "14";
-constexpr std::string_view VersionPatch = "0";
-constexpr std::string_view SpecVersion  = "1.10.4";
+constexpr std::string_view VersionPatch = "1";
+constexpr std::string_view SpecVersion  = "1.10.12";
 
 enum class Severity { INFO, WARNING, ERROR };